Menambahkan menu submenu sasaran pada menu master

+++ resources/views/Mst/sasaran-list.blade.php
@@ -1,6 +1,9 @@
 @extends('layouts.app')
 @section('content')
-
+<style media="screen">
+  .modal-lg{ width: 100% !important; }
+  .ui-datepicker{ z-index:99999 !important; }
+</style>
 <div class="br-pageheader pd-y-15 pd-l-20">
   <nav class="breadcrumb pd-0 mg-0 tx-12">
     <a class="breadcrumb-item" href="/">Dashboard</a>
@@ -35,7 +38,7 @@
           <div class="float-right">
 
             @if(can_access("mst_skpd", "add"))
-            <a class='btn btn-sm btn-success' href='{{url()->current()}}/add'><i class='menu-item-icon icon ion-plus'></i> Tambah</a>
+            <button type="button" class="btn btn-sm btn-success" data-toggle="modal" data-target="#addModal">Tambah</button>
             @endif
           </div>
         </div>
@@ -43,8 +46,11 @@
           <table class="table table-bordered table-striped responsive" id="oTable" style="width:100%">
             <thead>
               <tr>
-                <th>Nama Sasaran</th>
-                <th>Parent</th>
+                <th>Kegiatan</th>
+                <th>Sasaran</th>
+                <th>Perangkat Daerah</th>
+                <th>Dari</th>
+                <th>Sampai</th>
                 <th style='width:150px'>Aksi</th>
               </tr>
             </thead>
@@ -54,9 +60,18 @@
     </div>
   </div>
 </div>
+<!-- modal add -->
+@include('Mst.sasaran-form_add')
+
+<!-- modal edit -->
+@include('Mst.sasaran-form_edit')
+
 @endsection
 
 @section('scripts')
+<!-- Date range picker -->
+<!-- <script type="text/javascript" src="{{ asset('admin_template/lib/daterangepicker/moment.min.js') }}"></script>
+<script type="text/javascript" src="{{ asset('admin_template/lib/daterangepicker/daterangepicker.js') }}"></script> -->
 <!-- Datatables -->
 <script src="{{ asset('admin_template/lib/datatables/jquery.dataTables.js') }}"></script>
 <script>
@@ -72,22 +87,31 @@ $(function() {
       serverSide: true,
       ajax: '{{url()->current()}}/datatables/',
       columns: [
-        { data: 'nama', name: 's.nama'},
-        { data: 'nama_parent', name: 'p.nama'},
+        { data: 'kegiatan', name: 'kegiatan'},
+        { data: 'skpd', name: 'skpd'}, // sasaran
+        { data: 'skpd', name: 'skpd'},
+        { data: 'dari', name:'dari', orderable: false, render: function ( data, type, row ) {
+          var x = new Date(data);
+          return moment(x).format("DD-MM-YYYY");
+        }},
+        { data: 'sampai', name:'sampai', orderable: false, render: function ( data, type, row ) {
+          var x = new Date(data);
+          return moment(x).format("DD-MM-YYYY");
+        }},
         { data: null, name:null, orderable: false, render: function ( data, type, row ) {
           var return_button = "";
           @if(can_access("mst_skpd", "edit"))
-          return_button += "<a class='btn btn-warning btn-xs' href='{{url()->current()}}/edit/" + data.id + "'><i class='fa fa-pencil'></i> Edit</a> ";
+          return_button += "<button class='btn btn-warning btn-xs' data-toggle='modal' data-target='#editModal' data-id='" + data.id + "'><i class='fa fa-pencil'></i> Edit</button> ";
           @endif
           @if(can_access("mst_skpd", "delete"))
-          return_button += "<a class='btn btn-danger btn-xs' href='{{url()->current()}}/delete/" + data.id + "'><i class='fa fa-close'></i> Hapus</a>";
+          return_button += "<a class='btn btn-danger btn-xs' href='{{url()->current()}}/delete/" + data.id + "' onclick='return confirm(\"Apakah anda ingin menghapus data ini?\")'><i class='fa fa-close'></i> Hapus</a>";
           @endif
           return return_button == "" ? "-" : return_button;
         }},
       ],
       columnDefs: [
       {
-        targets: 2,
+        targets: 5,
         className: "text-center",
      }],
   });
@@ -96,6 +120,12 @@ $(function() {
   setTimeout(function() {
     $(".alert-success").hide(1000);
   }, 3000);
+
+  $('#addModal, #editModal').on('show.bs.modal', function () {
+    $(this).find('form').trigger('reset');
+    $("#cover-sasaran").html('');
+    $("#cover-sasaran_edit").html('');
+  });
 
 });
 </script>
